+++ src/parsers/contest/KattisContestParser.ts
@@ -2,7 +2,7 @@ import { KattisProblemParser } from '../problem/KattisProblemParser';
 import { SimpleContestParser } from '../SimpleContestParser';
 
 export class KattisContestParser extends SimpleContestParser {
-  protected linkSelector = '#standings > thead > tr > th.problemcolheader-standings > a';
+  protected linkSelector = 'table.table2 > tbody > tr > td:first-of-type > a';
   protected problemParser = new KattisProblemParser();
 
   public getMatchPatterns(): string[] {
@@ -10,6 +10,6 @@ export class KattisContestParser extends SimpleContestParser {
   }
 
   public getRegularExpressions(): RegExp[] {
-    return [/^https:\/\/.*[.]kattis[.]com\/contests\/([a-z0-9]+)(\/standings)?$/];
+    return [/^https:\/\/.*[.]kattis[.]com\/contests\/([a-z0-9]+)(\/problems)?$/];
   }
 }