+++ src/parsers/problem/KattisProblemParser.ts
@@ -18,16 +18,15 @@ export class KattisProblemParser extends Parser {
 
     task.setName(elem.querySelector('h1').innerHTML.replace(/<br>/, ' - '));
 
-    const contestNode = elem.querySelector('h2.title');
-    if (contestNode !== null) {
-      task.setCategory(contestNode.textContent);
+    if (elem.querySelector('.breadcrumb > .breadcrumb-link:first-child').textContent === 'Contests') {
+      task.setCategory(elem.querySelector('.breadcrumb > .breadcrumb-link:nth-of-type(2)').textContent);
     }
 
     task.setInteractive([...elem.querySelectorAll('h2')].some(el => (el as any).textContent === 'Interaction'));
 
-    const sidebar = elem.querySelector('.problem-sidebar').textContent;
-    task.setTimeLimit(parseFloat(/([0-9.]+) second/.exec(sidebar)[1]) * 1000);
-    task.setMemoryLimit(parseInt(/(\d+) MB/.exec(sidebar)[1], 10));
+    const limitsStr = elem.querySelector('.attribute_list-book.attribute_list').textContent;
+    task.setTimeLimit(parseFloat(/([0-9.]+) second/.exec(limitsStr)[1]) * 1000);
+    task.setMemoryLimit(parseInt(/(\d+) MB/.exec(limitsStr)[1], 10));
 
     elem.querySelectorAll('.sample').forEach(table => {
       const blocks = table.querySelectorAll('pre');